20230719- 44
Các bug còn lại
1. Tập trung hoàn thành biểu đồ và phân tích nhiều thành tố vào
2. Mấy chỗ EditText quá nguy hiểm. Cần 1 chuyên gia TestCase ở đấy
3. Giải quyết triệt để thằng customer
4. Testcase nhiều vào cho các màn hình

+++ app/src/main/java/com/example/restaurantpos/db/entity/BillEntity.kt
@@ -36,7 +36,7 @@ data class BillEntity constructor(
     var bill_customer_rank_percent: Int,
 
 //    @ColumnInfo(name = "bill_tax")
-    var bill_tax: Int,
+    var bill_tax: Float,
 
 //    @ColumnInfo(name = "bill_total")
     var bill_total: Float,
